divide transaction editor into two components for edit and review
