+++ app/components/TransactionEdit/index.js
@@ -1,0 +1,314 @@
+import React, { Component } from 'react';
+import AccountInput from '../AccountInput';
+import AccountSelect from '../AccountSelect';
+import ActionView from '../ActionView';
+import EthInput from '../EthInput';
+import PropTypes from 'prop-types';
+import { StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
+import { colors, fontStyles } from '../../styles/common';
+import { connect } from 'react-redux';
+import { toBN, isBN, hexToBN } from '../../util/number';
+import { isValidAddress, toChecksumAddress } from 'ethereumjs-util';
+import { strings } from '../../../locales/i18n';
+import { withNavigation } from 'react-navigation';
+import CustomGas from '../CustomGas';
+
+const styles = StyleSheet.create({
+	root: {
+		backgroundColor: colors.white,
+		flex: 1
+	},
+	formRow: {
+		flexDirection: 'row',
+		height: 100
+	},
+	fromRow: {
+		zIndex: 5
+	},
+	toRow: {
+		marginTop: 16,
+		zIndex: 4
+	},
+	amountRow: {
+		marginTop: 16,
+		zIndex: 3
+	},
+	label: {
+		flex: 0,
+		paddingRight: 18,
+		width: 96
+	},
+	labelText: {
+		...fontStyles.bold,
+		color: colors.gray,
+		fontSize: 16
+	},
+	max: {
+		...fontStyles.bold,
+		color: colors.primary,
+		fontSize: 12,
+		paddingTop: 6
+	},
+	error: {
+		...fontStyles.bold,
+		color: colors.red,
+		fontSize: 12,
+		lineHeight: 12,
+		paddingTop: 6
+	},
+	form: {
+		flex: 1,
+		padding: 16
+	},
+	hexData: {
+		...fontStyles.bold,
+		backgroundColor: colors.white,
+		borderColor: colors.inputBorderColor,
+		borderRadius: 4,
+		borderWidth: 1,
+		flex: 1,
+		fontSize: 16,
+		minHeight: 64,
+		paddingLeft: 10,
+		paddingVertical: 6
+	}
+});
+
+/**
+ * Component that supports editing and reviewing a transaction
+ */
+class TransactionEdit extends Component {
+	static propTypes = {
+		/**
+		 * List of accounts from the AccountTrackerController
+		 */
+		accounts: PropTypes.object,
+		/**
+		 * react-navigation object used for switching between screens
+		 */
+		navigation: PropTypes.object,
+		/**
+		 * Hids the "data" field
+		 */
+		hideData: PropTypes.bool,
+		/**
+		 * Callback triggered when this transaction is cancelled
+		 */
+		onCancel: PropTypes.func,
+		/**
+		 * Called when a user changes modes
+		 */
+		onModeChange: PropTypes.func,
+		/**
+		 * Currently-active account address in the current keychain
+		 */
+		selectedAddress: PropTypes.string,
+		/**
+		 * Transaction object associated with this transaction
+		 */
+		transaction: PropTypes.object,
+		/**
+		 * Callback to open the qr scanner
+		 */
+		onScanSuccess: PropTypes.func,
+		/**
+		 * Callback to update amount in parent state
+		 */
+		handleUpdateAmount: PropTypes.func,
+		/**
+		 * Callback to update gas and gasPrice in parent state
+		 */
+		handleGasFeeSelection: PropTypes.func,
+		/**
+		 * Amount BN object containing amount selected for transaction
+		 */
+		amount: PropTypes.string,
+		/**
+		 * Amount BN object containing gas selected for transaction
+		 */
+		gas: PropTypes.object,
+		/**
+		 * Amount BN object containing gasPrice selected for transaction
+		 */
+		gasPrice: PropTypes.object
+	};
+
+	state = {
+		data: this.props.transaction.data,
+		from: this.props.transaction.from,
+		to: this.props.transaction.to,
+		toFocused: false
+	};
+
+	fillMax = () => {
+		const { gas, gasPrice } = this.props;
+		const { balance } = this.props.accounts[this.state.from];
+		this.setState({});
+		this.props.handleUpdateAmount({
+			amount: !isBN(gas) || !isBN(gasPrice) ? hexToBN(balance) : hexToBN(balance).sub(gas.mul(gasPrice))
+		});
+	};
+
+	onFocusToAddress = () => {
+		this.setState({ toFocused: true });
+	};
+
+	onCancel = () => {
+		const { onCancel } = this.props;
+		onCancel && onCancel();
+	};
+
+	review = async () => {
+		const { onModeChange } = this.props;
+		await this.setState({ toFocused: true });
+		!this.validate() && onModeChange && onModeChange('review');
+	};
+
+	updateAmount = amount => {
+		this.props.handleUpdateAmount(amount);
+	};
+
+	updateData = data => {
+		this.setState({ data });
+	};
+
+	updateFromAddress = from => {
+		this.setState({ from });
+	};
+
+	updateToAddress = to => {
+		this.setState({ to });
+	};
+
+	validate() {
+		if (this.validateAmount() || this.validateGas() || this.validateToAddress()) {
+			return true;
+		}
+	}
+
+	validateAmount() {
+		let error;
+		const { from } = this.state;
+		const { amount, gas, gasPrice } = this.props;
+		const checksummedFrom = from ? toChecksumAddress(from) : '';
+		const fromAccount = this.props.accounts[checksummedFrom];
+		amount && !isBN(amount) && (error = strings('transaction.invalidAmount'));
+		amount &&
+			fromAccount &&
+			isBN(gas) &&
+			isBN(gasPrice) &&
+			isBN(amount) &&
+			hexToBN(fromAccount.balance).lt(amount.add(gas.mul(gasPrice))) &&
+			(error = strings('transaction.insufficient'));
+		return error;
+	}
+
+	validateGas() {
+		let error;
+		const { gas, gasPrice } = this.props;
+		gas && !isBN(gas) && (error = strings('transaction.invalidGas'));
+		gasPrice && !isBN(gasPrice) && (error = strings('transaction.invalidGasPrice'));
+		return error;
+	}
+
+	validateToAddress() {
+		let error;
+		const { to } = this.state;
+		!to && this.state.toFocused && (error = strings('transaction.required'));
+		to && !isValidAddress(to) && (error = strings('transaction.invalidAddress'));
+		return error;
+	}
+
+	onScanSuccess = () => {
+		this.props.navigation.navigate('QRScanner', {
+			onScanSuccess: this.props.onScanSuccess,
+			addressOnly: true
+		});
+	};
+
+	render() {
+		const { data, from = this.props.selectedAddress, to } = this.state;
+		const { hideData, amount, gas, gasPrice } = this.props;
+		const totalGas = isBN(gas) && isBN(gasPrice) ? gas.mul(gasPrice) : toBN('0x0');
+
+		return (
+			<View style={styles.root}>
+				<ActionView confirmText="Next" onCancelPress={this.onCancel} onConfirmPress={this.review}>
+					<View style={styles.form}>
+						<View style={{ ...styles.formRow, ...styles.fromRow }}>
+							<View style={styles.label}>
+								<Text style={styles.labelText}>{strings('transaction.from')}:</Text>
+							</View>
+							<AccountSelect value={from} onChange={this.updateFromAddress} />
+						</View>
+						<View style={{ ...styles.formRow, ...styles.toRow }}>
+							<View style={styles.label}>
+								<Text style={styles.labelText}>{strings('transaction.to')}:</Text>
+								{this.validateToAddress() && (
+									<Text style={styles.error}>{this.validateToAddress()}</Text>
+								)}
+							</View>
+							<AccountInput
+								onChange={this.updateToAddress}
+								onFocus={this.onFocusToAddress}
+								placeholder={strings('transaction.recipientAddress')}
+								onScanSuccess={this.onScanSuccess}
+								value={to}
+							/>
+						</View>
+						<View style={{ ...styles.formRow, ...styles.amountRow }}>
+							<View style={styles.label}>
+								<Text style={styles.labelText}>{strings('transaction.amount')}:</Text>
+								{this.validateAmount() ? (
+									<Text style={styles.error}>{this.validateAmount()}</Text>
+								) : (
+									<TouchableOpacity onPress={this.fillMax}>
+										<Text style={styles.max}>{strings('transaction.max')}</Text>
+									</TouchableOpacity>
+								)}
+							</View>
+							<EthInput onChange={this.updateAmount} value={amount} />
+						</View>
+						<View style={{ ...styles.formRow, ...styles.amountRow }}>
+							<View style={styles.label}>
+								<Text style={styles.labelText}>{strings('transaction.gasFee')}:</Text>
+								{this.validateGas() && <Text style={styles.error}>{this.validateGas()}</Text>}
+							</View>
+							<EthInput readonly value={totalGas} />
+						</View>
+						{!hideData && (
+							<View style={{ ...styles.formRow, ...styles.amountRow }}>
+								<View style={styles.label}>
+									<Text style={styles.labelText}>{strings('transaction.hexData')}:</Text>
+								</View>
+								<TextInput
+									multiline
+									onChangeText={this.updateData}
+									placeholder="Optional"
+									style={styles.hexData}
+									value={data}
+								/>
+							</View>
+						)}
+						<View style={{ ...styles.formRow, ...styles.amountRow }}>
+							<View style={styles.label}>
+								<Text style={styles.labelText}>{strings('transaction.gasFee')}:</Text>
+							</View>
+							<CustomGas
+								transaction={this.props.transaction}
+								handleGasFeeSelection={this.props.handleGasFeeSelection}
+							/>
+						</View>
+					</View>
+				</ActionView>
+			</View>
+		);
+	}
+}
+
+const mapStateToProps = state => ({
+	accounts: state.backgroundState.AccountTrackerController.accounts,
+	selectedAddress: state.backgroundState.PreferencesController.selectedAddress
+});
+
+export default withNavigation(connect(mapStateToProps)(TransactionEdit));
